Move check from `SurfaceVertex` to `HalfEdge`
I'm working on removing the reference to `Surface` from `SurfaceVertex`.
Once that is gone, it won't be able to support that validation check any
longer.
